update index and create states module
